requirements.txt
gradio==4.15.0
gradio_client==0.8.1
openai==1.13.3
python-dotenv==1.0.1
